+++ .c8rc
@@ -4,10 +4,8 @@
     "lcov"
   ],
   "exclude": [
-    "ava.config.js",
     "**/test/**",
     "**/tools/**",
-    "**/node_modules/**",
-    ".hostlocal.js"
+    "**/node_modules/**"
   ]
 }
