--- ava.config.js
@@ -1,5 +1,0 @@
-export default {
-  files: [
-    'test/*.ava.js',
-  ],
-};
