+++ package.json
@@ -10,7 +10,7 @@
     "build": "tsup",
     "lint": "eslint .",
     "pretest": "npm run build -- --sourcemap inline --no-minify --no-dts",
-    "test": "c8 ava",
+    "test": "c8 node --test test/*.test.js",
     "posttest": "tsup --silent",
     "docs": "typedoc",
     "clean": "rimraf docs lib coverage src/grammar.js src/grammar.d.ts"
@@ -38,7 +38,6 @@
     "@types/node": "22.13.9",
     "@typescript-eslint/eslint-plugin": "8.26.0",
     "@typescript-eslint/parser": "8.26.0",
-    "ava": "6.2.0",
     "c8": "10.1.3",
     "eslint": "9.21.0",
     "eslint-plugin-jsdoc": "50.6.3",
@@ -52,13 +51,13 @@
     "typescript": "5.8.2",
     "typescript-eslint": "8.26.0"
   },
-  "packageManager": "pnpm@10.6.0",
-  "engines": {
-    "node": ">=18"
-  },
   "pnpm": {
     "onlyBuiltDependencies": [
       "esbuild"
     ]
+  },
+  "packageManager": "pnpm@10.6.0",
+  "engines": {
+    "node": ">=18.8"
   }
 }
