--- test/utils.ava.js
@@ -1,6 +1,0 @@
-import {asciiToUint8Array} from '../lib/utils.js';
-import test from 'ava';
-
-test('asciiToUint8Array', t => {
-  t.throws(() => asciiToUint8Array('\u{256}'));
-});
